# [go/mysql] use tiered pool for buffers to avoid false hits