Now when we get smaller than we want buffer from pool, we just put it
back and allocate a new one. This might be quite inefficient.
So, instead use multiple pools which always return buffers >= request
size.
Benchmark results:
```
benchmark                            old ns/op     new ns/op     delta
BenchmarkParallelShortQueries-8      3347          3327          -0.60%
BenchmarkParallelMediumQueries-8     6535          5789          -11.42%
BenchmarkParallelRandomQueries-8     6001539       5453566       -9.13%

benchmark                            old MB/s     new MB/s     speedup
BenchmarkParallelShortQueries-8      3.29         3.31         1.01x
BenchmarkParallelMediumQueries-8     2507.72      2830.78      1.13x

benchmark                            old allocs     new allocs     delta
BenchmarkParallelShortQueries-8      23             23             +0.00%
BenchmarkParallelMediumQueries-8     9              8              -11.11%
BenchmarkParallelRandomQueries-8     15             14             -6.67%

benchmark                            old bytes     new bytes     delta
BenchmarkParallelShortQueries-8      720           720           +0.00%
BenchmarkParallelMediumQueries-8     27243         20865         -23.41%
BenchmarkParallelRandomQueries-8     36496086      29452887      -19.30%
```
This is new package and internal to vitess. So, we can make interface less flexible maybe. Like allow only powers of two and say that multiplier is always 2.