@sougou 2018-09-05
Let's use a human-readable simplification:
minSize: 1
maxSize: 16
multiipler: 2

This will yield the following buckets: 1, 2, 4, 8, 16.

Now, if we perform log operations, with multiplier as base(2), it will yield the index of the pool.
The respective logs for the above will be: 0, 1, 2, 3, 4.

Let's say requested length was: 5. Log(5) = 2.3, which rounds up to 3.
Log(9) -> 3.1, rounds up to 4.

This approach will allow us to have a larger number of pools. We should have pools of size up to 16M.
Hopefully, log functions aren't too expensive :).

For minSize other than 1, we have to first divide by minSize before taking the log.