@sougou 2018-09-05
As for returning the last index, the idea is to put everything above the max length into that bucket and take a gamble. The next one that pulls from the pool will have to check the length.

But I think that's not worth it. So, it's better to just allocate from heap.