@LK4D4 2018-09-05
@sougou is it okay to fix multiplier as 2? Because there is no arbitrary base logarithm in go it seems.