@sougou 2018-09-05
Yeah. I was surprised about that. But 2 is a good multiplier to hard-code.
Let's also run this benchmark to see how fast it goes: https://golang.org/src/math/all_test.go#L3404 :)